Modular unit creation
Added the ability to easily develop units. Developer can modify all aspects of a unit, create their own animation libraries, sprites and give them unique abilities without causing issues with other entities. Thank you lord of OOP!

Placeholders for buildings added for the same process but haven't created the necessary node structure for it as of yet.

Chuffed with this